+++ lib/frontend/src/components/IssueDetail.jsx
@@ -9,7 +9,8 @@ import {
   updateIssueLabels,
 } from '../actions/issues'
 import { block, connect } from '../utils'
-import LabelMultiSelect from './LabelMultiSelect'
+import IssueStatusIcon from './Utils/IssueStatusIcon'
+import LabelMultiSelect from './Utils/LabelMultiSelect'
 
 const b = block('IssueDetail')
 var ReactMarkdown = require('react-markdown')
@@ -99,33 +100,7 @@ class IssueDetail extends React.Component {
             </span>
           </span>
           <span className={b('infos')}>
-            {issue.pull_request === null ? (
-              issue.state === 'closed' ? (
-                <Octicon
-                  className="githubIcons"
-                  name="issue-closed"
-                  style={{ color: 'red' }}
-                />
-              ) : (
-                <Octicon
-                  className="githubIcons"
-                  name="issue-opened"
-                  style={{ color: 'green' }}
-                />
-              )
-            ) : issue.state === 'closed' ? (
-              <Octicon
-                className="githubIcons"
-                name="git-merge"
-                style={{ color: '#6f42c1' }}
-              />
-            ) : (
-              <Octicon
-                className="githubIcons"
-                name="git-pull-request"
-                style={{ color: 'green' }}
-              />
-            )}
+            <IssueStatusIcon issue={issue} />
             {issue.user.user_name} opened this issue, last update:{' '}
             {format(new Date(issue.updated_at), 'DD/MM/YYYY HH:mm:ss')}
           </span>
